Avoid unnecessary allocations
Instead of collecting generated words into a Vec, and then building the
passphrase String our of that, allocate a String beforehand and add
generated words and delimiters to it. That way there's only ever one
allocation. Since the String is allocated with a worst-case size, there
will never be a need for a re-allocation.